Merge remote-tracking branch 'OAGithubRStudio/main' into split_by-numeric

# Conflicts:
#	R/gitbook.R
#	R/html.R
#	man/build_chapter.Rd
#	man/epub_book.Rd
#	man/gitbook.Rd
#	man/html_chapters.Rd
#	man/html_document2.Rd
#	man/pdf_book.Rd
#	man/publish_book.Rd
#	man/render_book.Rd
#	man/resolve_refs_html.Rd
#	man/serve_book.Rd

+++ R/render.R
@@ -9,15 +9,20 @@
 #' (\code{new_session = FALSE}) is to merge Rmd files into a single file and
 #' render this file. You can also choose to render each individual Rmd file in a
 #' new R session (\code{new_session = TRUE}).
-#' @param input An input filename (or multiple filenames). If \code{preview =
-#'   TRUE}, only files specified in this argument are rendered, otherwise all R
-#'   Markdown files specified by the book are rendered.
+#' @param input A directory, an input filename or multiple filenames. For a
+#'   directory, \file{index.Rmd} will be used if it exists in this (book)
+#'   project directory. For filenames, if \code{preview = TRUE}, only files
+#'   specified in this argument are rendered, otherwise all R Markdown files
+#'   specified by the book are rendered.
 #' @param output_format,...,clean,envir Arguments to be passed to
-#'   \code{rmarkdown::\link[rmarkdown]{render}()}. For \code{preview_chapter()},
-#'   \code{...} is passed to \code{render_book()}.
-#' @param clean_envir Whether to clean up the environment \code{envir} before
-#'   rendering the book. By default, the environment is cleaned when rendering
-#'   the book in a non-interactive R session.
+#'   \code{rmarkdown::\link{render}()}. For \code{preview_chapter()}, \code{...}
+#'   is passed to \code{render_book()}. See \code{rmarkdown::\link{render}()}
+#'   and \href{https://bookdown.org/yihui/bookdown/build-the-book.html}{the
+#'   bookdown reference book} for details on how output formatting options are
+#'   set from YAML or parameters supplied by the user when calling
+#'   \code{render_book()}.
+#' @param clean_envir This argument has been deprecated and will be removed in
+#'   future versions of \pkg{bookdown}.
 #' @param output_dir The output directory. If \code{NULL}, a field named
 #'   \code{output_dir} in the configuration file \file{_bookdown.yml} will be
 #'   used (possibly not specified, either, in which case a directory name
@@ -30,38 +35,69 @@
 #'   \code{input} argument. Previewing a certain chapter may save compilation
 #'   time as you actively work on this chapter, but the output may not be
 #'   accurate (e.g. cross-references to other chapters will not work).
-#' @param encoding Ignored. The character encoding of all input files is
-#'   supposed to be UTF-8.
 #' @param config_file The book configuration file.
 #' @export
 #' @examples
 #' # see https://bookdown.org/yihui/bookdown for the full documentation
 #' if (file.exists('index.Rmd')) bookdown::render_book('index.Rmd')
+#' \dontrun{
+#' # will use the default format defined in index.Rmd or _output.yml
+#' bookdown::render_book("index.Rmd")
+#' # will use the options for format defined in YAML metadata
+#' bookdown::render_book("index.Rmd",  "bookdown::pdf_book")
+#' # If you pass an output format object, it must have all the options set
+#' bookdown::render_book("index.Rmd", bookdown::pdf_book(toc = FALSE))
+#'
+#' # will render the book in the current directory
+#' bookdown::render_book()
+#' # this is equivalent to
+#' bookdown::render_book("index.Rmd")
+#' # will render the book living in the specified directory
+#' bookdown::render_book("my_book_project")
+#' }
 render_book = function(
-  input, output_format = NULL, ..., clean = TRUE, envir = parent.frame(),
+  input = ".", output_format = NULL, ..., clean = TRUE, envir = parent.frame(),
   clean_envir = !interactive(), output_dir = NULL, new_session = NA,
-  preview = FALSE, encoding = 'UTF-8', config_file = '_bookdown.yml'
+  preview = FALSE, config_file = '_bookdown.yml'
 ) {
 
   verify_rstudio_version()
+
+  # select and check input file(s)
+  if (length(input) == 1L && dir_exists(input)) {
+    message(sprintf("Rendering book in directory '%s'", input))
+    owd = setwd(input); on.exit(setwd(owd), add = TRUE)
+    input = "index.Rmd"
+  }
+
+  stop_if_not_exists(input)
+
   format = NULL  # latex or html
   if (is.list(output_format)) {
     format = output_format$bookdown_output_format
     if (!is.character(format) || !(format %in% c('latex', 'html'))) format = NULL
-  } else if (is.character(output_format)) {
-    if (identical(output_format, 'all')) {
-      output_format = rmarkdown::all_output_formats(input, 'UTF-8')
+  } else if (is.null(output_format) || is.character(output_format)) {
+    if (is.null(output_format) || identical(output_format, 'all')) {
+      # formats can safely be guess when considering index.Rmd and its expected frontmatter
+      # and not another Rmd file which has no expected YAML frontmatter
+      stop_if_not_exists("index.Rmd")
+      all_formats = rmarkdown::all_output_formats("index.Rmd")
+      # when no format provided, return name of the first resolved
+      output_format = if (is.null(output_format)) all_formats[[1]] else all_formats
     }
-    if (length(output_format) > 1) {
-      return(unlist(lapply(output_format, function(fmt) render_book(
+    if (length(output_format) > 1) return(unlist(lapply(output_format, function(fmt)
+      xfun::Rscript_call(render_book, list(
         input, fmt, ..., clean = clean, envir = envir, output_dir = output_dir,
         new_session = new_session, preview = preview, config_file = config_file
-      ))))
-    }
+      ), fail = c("bookdown::render_book() failed to render the output format '", fmt, "'."))
+    )))
     format = target_format(output_format)
   }
 
-  if (clean_envir) rm(list = ls(envir, all.names = TRUE), envir = envir)
+  if (!missing(clean_envir)) warning(
+    "The argument 'clean_envir' has been deprecated and will be removed in future ",
+    "versions of bookdown."
+  )
 
   if (config_file != '_bookdown.yml') {
     unlink(tmp_config <- tempfile('_bookdown_', '.', '.yml'))
@@ -76,17 +112,16 @@ render_book = function(
   on.exit(opts$restore(), add = TRUE)
   config = load_config()  # configurations in _bookdown.yml
   output_dir = output_dirname(output_dir, config)
-  on.exit(clean_empty_dir(output_dir), add = TRUE)
+  on.exit(xfun::del_empty_dir(output_dir), add = TRUE)
+  if (!preview) unlink(ref_keys_path(output_dir))  # clean up reference-keys.txt
   # store output directory and the initial input Rmd name
   opts$set(
-    output_dir = output_dir, input_rmd = basename(input), preview = preview
+    output_dir = output_dir,
+    input_rmd = xfun::relative_path(input),
+    preview = preview
   )
 
   aux_diro = '_bookdown_files'
-  # for compatibility with bookdown <= 0.0.64
-  if (isTRUE(dir_exists(aux_dir2 <- file.path(output_dir, aux_diro)))) {
-    if (!dir_exists(aux_diro)) file.rename(aux_dir2, aux_diro)
-  }
   # move _files and _cache from _bookdown_files to ./, then from ./ to _bookdown_files
   aux_dirs = files_cache_dirs(aux_diro)
   move_dirs(aux_dirs, basename(aux_dirs))
@@ -106,24 +141,27 @@ render_book = function(
 
   main = book_filename()
   if (!grepl('[.][Rr]?md$', main)) main = paste0(main, if (new_session) '.md' else '.Rmd')
-  delete_main = isTRUE(config[['delete_merged_file']])
-  if (file.exists(main) && !delete_main) stop(
-    'The file ', main, ' exists. Please delete it if it was automatically generated, ',
-    'or set a different book_filename option in _bookdown.yml. If you are sure ',
-    "it can be safely deleted, please set the option 'delete_merged_file' to true in _bookdown.yml."
+  delete_main = config[['delete_merged_file']]
+  check_main = function() file.exists(main) && is.null(delete_main)
+  if (check_main()) stop(
+    'The file ', main, ' exists. Please delete it if it was automatically generated. ',
+    'If you are sure it can be safely overwritten or deleted, please set the option ',
+    "'delete_merged_file' to true in _bookdown.yml."
   )
-  on.exit(if (file.exists(main) && !delete_main) {
+  on.exit(if (check_main()) {
     message('Please delete ', main, ' after you finish debugging the error.')
   }, add = TRUE)
   opts$set(book_filename = main)  # store the book filename
 
-  files = setdiff(source_files(format, config), main)
+  files = source_files(format, config)
   if (length(files) == 0) stop(
     'No input R Markdown files found from the current directory ', getwd(),
     ' or in the rmd_files field of _bookdown.yml'
   )
   if (new_session && any(dirname(files) != '.')) stop(
-    'All input files must be under the current working directory'
+    'With new_session = TRUE, all input files must be under the root directory ',
+    'of the (book) project. You might have used `rmd_files` or `rmd_subdir` to ',
+    'specify input files from subdirectories, which will not work with `new_session`.'
   )
 
   res = if (new_session) {
@@ -131,7 +169,7 @@ render_book = function(
   } else {
     render_cur_session(files, main, config, output_format, clean, envir, ...)
   }
-  file.remove(main)
+  if (!xfun::isFALSE(delete_main)) file.remove(main)
   res
 }
 
@@ -147,7 +185,7 @@ render_cur_session = function(files, main, config, output_format, clean, envir, 
     insert_chapter_script(config, 'before'),
     insert_chapter_script(config, 'after')
   )
-  rmarkdown::render(main, output_format, ..., clean = clean, envir = envir, encoding = 'UTF-8')
+  rmarkdown::render(main, output_format, ..., clean = clean, envir = envir)
 }
 
 render_new_session = function(files, main, config, output_format, clean, envir, ...) {
@@ -169,26 +207,12 @@ render_new_session = function(files, main, config, output_format, clean, envir, 
   # if input is index.Rmd or not preview mode, compile all Rmd's
   rerun = !opts$get('preview') || identical(opts$get('input_rmd'), 'index.Rmd')
   if (!rerun) rerun = files %in% opts$get('input_rmd')
-  add1 = insert_chapter_script(config, 'before')
-  add2 = insert_chapter_script(config, 'after')
+  add1 = merge_chapter_script(config, 'before')
+  add2 = merge_chapter_script(config, 'after')
+  on.exit(unlink(c(add1, add2)), add = TRUE)
   # compile chapters in separate R sessions
-  for (f in files[rerun]) {
-    if (length(add1) + length(add2) == 0) {
-      Rscript_render(f, render_args, render_meta)
-      next
-    }
-    # first backup the original Rmd to a tempfile
-    f2 = tempfile('bookdown', '.', fileext = '.bak')
-    file.copy(f, f2, overwrite = TRUE)
-    # write add1/add2 to the original Rmd, compile it, and restore it
-    tryCatch({
-      txt = c(add1, read_utf8(f), add2)
-      write_utf8(txt, f)
-      Rscript_render(f, render_args, render_meta)
-    }, finally = {
-      if (file.copy(f2, f, overwrite = TRUE)) file.remove(f2)
-    })
-  }
+  for (f in files[rerun]) Rscript_render(f, render_args, render_meta, add1, add2)
+
   if (!all(dirname(files_md) == '.'))
     file.copy(files_md[!rerun], basename(files_md[!rerun]), overwrite = TRUE)
 
@@ -204,7 +228,7 @@ render_new_session = function(files, main, config, output_format, clean, envir, 
 
   rmarkdown::render(
     main, output_format, ..., clean = clean, envir = envir,
-    run_pandoc = TRUE, knit_meta = knit_meta, encoding = 'UTF-8'
+    run_pandoc = TRUE, knit_meta = knit_meta
   )
 
 }
